Various structural cleanups of node-driver: * Extracted out ShutdownManager into its own file * Moved RPCDriver and ProcessUtilities into internal package * Made n.c.testing.performance package internal
