+++ testing/node-driver/src/integration-test/kotlin/net/corda/testing/internal/ProcessUtilitiesTests.kt
@@ -1,4 +1,4 @@
-package net.corda.testing.driver
+package net.corda.testing.internal
 
 import org.apache.commons.io.FileUtils
 import org.junit.Rule
