+++ testing/node-driver/src/main/kotlin/net/corda/testing/NodeTestUtils.kt
@@ -45,7 +45,6 @@ fun ledger(
  */
 @JvmOverloads
 fun transaction(
-        transactionLabel: String? = null,
         transactionBuilder: TransactionBuilder = TransactionBuilder(notary = DUMMY_NOTARY),
         initialiseSerialization: Boolean = true,
         cordappPackages: List<String> = emptyList(),
