+++ testing/node-driver/src/main/kotlin/net/corda/testing/internal/performance/Injectors.kt
@@ -1,4 +1,4 @@
-package net.corda.testing.performance
+package net.corda.testing.internal.performance
 
 import com.codahale.metrics.Gauge
 import com.codahale.metrics.MetricRegistry
